requirements.txt
PyQt5
pygame
requests
websockets
lxml
hidapi